feat: Regenerate latest version of the client with field presence support (#401)
Also add 7012 and 6003 backward compatibility check errors to ignorelist

This is totally expected, since it is an alpha client and both of those errors (`6003` - public constant value change, and `7012` - addition of a method to an interface) are expected given the underlying change. Specifially the `optional` keyword was added to `compute.proto`, which caused addition of `has<FieldName>()` methods to proto stubs and respective builders (error 7012). The field numbers change (error `6003`) was caused by the rollback of the previous hack (https://github.com/googleapis/googleapis-discovery/pull/31) which, in its turn, was caused by a bug in protobuf.